# perf: optimize Explorer.Chain.Cache.Blocks