--- a/apps/explorer/lib/explorer/chain/cache/blocks.ex
+++ b/apps/explorer/lib/explorer/chain/cache/blocks.ex
@@ -39,6 +39,27 @@ defmodule Explorer.Chain.Cache.Blocks do
 
   def drop_nonconsensus(number) when not is_nil(number), do: drop_nonconsensus([number])
 
+  @doc """
+  Prepares a block for cache update by aggregating transaction metrics and clearing the transactions list.
+
+  This function ensures that all transaction-related statistics are calculated
+  and stored in the block's virtual fields before removing the transactions
+  list. This optimization allows the block to retain computed aggregate data
+  while eliminating the need to store redundant transaction references that
+  are maintained separately in the database.
+
+  The function first triggers transaction aggregation to compute metrics such
+  as transaction counts, total fees, burnt fees, and priority fees. After
+  aggregation is complete, the transactions list is cleared to reduce memory
+  usage.
+
+  ## Parameters
+  - `block`: A Block struct that may contain transactions to be aggregated
+
+  ## Returns
+  - Block struct with aggregated transaction metrics and an empty transactions list
+  """
+  @spec sanitize_before_update(Block.t()) :: Block.t()
   def sanitize_before_update(block) do
     block = Block.aggregate_transactions(block)