--- a/apps/explorer/lib/explorer/helper.ex
+++ b/apps/explorer/lib/explorer/helper.ex
@@ -487,7 +487,6 @@ defmodule Explorer.Helper do
   end
 
   @doc """
-  <<<<<<< HEAD
   Extracts the method ID from an ABI specification.
 
   ## Parameters