Closes #12382 

## Changelog
- Now blocks are aggregated before storing in cache, so cache takes less memory and responds faster
- Along with this ^ another performance problem was observed and fixed, `priority_fee` was calculated for each block separately and not using preloaded transactions:
```elixir
priority_fee = block.base_fee_per_gas && BlockPriorityFeeCount.fetch(block.hash)
```
This most of the times issued a separate database request, so loading the list of blocks could result in up to 50 additional database requests
- Added benchmark for `api/v2/blocks`, to run: 
```bash
cd apps/block_scout_web
mix run benchmarks/block_scout_web/controllers/api/v2/block_controller_benchmark.exs
```
requires the same envs as tests

<details>
<summary>Benchmark results</summary>

```
##### With input  0 blocks   0 transactions per block #####
Name                                                      ips        average  deviation         median         99th %
GET /api/v2/blocks first page (cached) (master)        1.87 K      535.62 μs    ±71.46%      479.71 μs     2719.54 μs
GET /api/v2/blocks first page (cached)                 1.83 K      546.79 μs    ±75.85%      457.85 μs     2899.30 μs

Comparison: 
GET /api/v2/blocks first page (cached) (master)        1.87 K
GET /api/v2/blocks first page (cached)                 1.83 K - 1.02x slower +11.18 μs

Memory usage statistics:

Name                                                    average  deviation         median         99th %
GET /api/v2/blocks first page (cached) (master)        79.70 KB     ±0.67%       79.88 KB       79.88 KB
GET /api/v2/blocks first page (cached)                 79.73 KB     ±0.69%       79.91 KB       79.91 KB

Comparison: 
GET /api/v2/blocks first page (cached) (master)        79.88 KB
GET /api/v2/blocks first page (cached)                 79.73 KB - 1.00x memory usage +0.0308 KB

##### With input 51 blocks   0 transactions per block #####
Name                                                      ips        average  deviation         median         99th %
GET /api/v2/blocks first page (cached) (master)        277.58        3.60 ms    ±27.60%        3.40 ms        7.99 ms
GET /api/v2/blocks first page (cached)                 262.76        3.81 ms    ±23.17%        3.64 ms        6.21 ms

Comparison: 
GET /api/v2/blocks first page (cached) (master)        277.58
GET /api/v2/blocks first page (cached)                 262.76 - 1.06x slower +0.20 ms

Memory usage statistics:

Name                                                    average  deviation         median         99th %
GET /api/v2/blocks first page (cached) (master)         4.09 MB     ±0.02%        4.09 MB        4.10 MB
GET /api/v2/blocks first page (cached)                  4.10 MB     ±0.00%        4.10 MB        4.10 MB

Comparison: 
GET /api/v2/blocks first page (cached) (master)         4.09 MB
GET /api/v2/blocks first page (cached)                  4.10 MB - 1.00x memory usage +0.00801 MB

##### With input 51 blocks  10 transactions per block #####
Name                                                      ips        average  deviation         median         99th %
GET /api/v2/blocks first page (cached)                 271.09        3.69 ms    ±94.28%        3.47 ms        6.03 ms
GET /api/v2/blocks first page (cached) (master)        241.99        4.13 ms    ±65.58%        4.02 ms        6.47 ms

Comparison: 
GET /api/v2/blocks first page (cached)                 271.09
GET /api/v2/blocks first page (cached) (master)        241.99 - 1.12x slower +0.44 ms

Memory usage statistics:

Name                                                    average  deviation         median         99th %
GET /api/v2/blocks first page (cached)                  4.15 MB     ±0.00%        4.15 MB        4.15 MB
GET /api/v2/blocks first page (cached) (master)         4.60 MB     ±0.01%        4.60 MB        4.60 MB

Comparison: 
GET /api/v2/blocks first page (cached)                  4.15 MB
GET /api/v2/blocks first page (cached) (master)         4.60 MB - 1.11x memory usage +0.45 MB

##### With input 51 blocks 100 transactions per block #####
Name                                                      ips        average  deviation         median         99th %
GET /api/v2/blocks first page (cached)                 295.99        3.38 ms    ±18.31%        3.29 ms        5.08 ms
GET /api/v2/blocks first page (cached) (master)         64.96       15.39 ms    ±27.60%       17.06 ms       24.50 ms

Comparison: 
GET /api/v2/blocks first page (cached)                 295.99
GET /api/v2/blocks first page (cached) (master)         64.96 - 4.56x slower +12.01 ms

Memory usage statistics:

Name                                                    average  deviation         median         99th %
GET /api/v2/blocks first page (cached)                  4.16 MB     ±0.00%        4.16 MB        4.16 MB
GET /api/v2/blocks first page (cached) (master)         8.61 MB     ±0.01%        8.61 MB        8.61 MB

Comparison: 
GET /api/v2/blocks first page (cached)                  4.16 MB
GET /api/v2/blocks first page (cached) (master)         8.61 MB - 2.07x memory usage +4.45 MB

##### With input 51 blocks 500 transactions per block #####
Name                                                      ips        average  deviation         median         99th %
GET /api/v2/blocks first page (cached)                 312.34        3.20 ms    ±16.99%        3.13 ms        4.76 ms
GET /api/v2/blocks first page (cached) (master)         12.38       80.77 ms    ±32.57%       86.77 ms      135.85 ms

Comparison: 
GET /api/v2/blocks first page (cached)                 312.34
GET /api/v2/blocks first page (cached) (master)         12.38 - 25.23x slower +77.56 ms

Memory usage statistics:

Name                                                    average  deviation         median         99th %
GET /api/v2/blocks first page (cached)                  4.15 MB     ±0.00%        4.15 MB        4.15 MB
GET /api/v2/blocks first page (cached) (master)        26.59 MB     ±0.00%       26.59 MB       26.59 MB

Comparison: 
GET /api/v2/blocks first page (cached)                  4.15 MB
GET /api/v2/blocks first page (cached) (master)        26.59 MB - 6.40x memory usage +22.43 MB
```
</details>
tldr: with quite heavy blocks (that have a lot of transactions) old cache can take up to 80.77 ms to respond while optimized version takes 3.20 ms


## Checklist for your Pull Request (PR)

- [ ] If I added new functionality, I added tests covering it.
- [ ] If I fixed a bug, I added a regression test to prevent the bug from silently reappearing again.
- [ ] I checked whether I should update the docs and did so by submitting a PR to [docs repository](https://github.com/blockscout/docs).
- [ ] If I added/changed/removed ENV var, I submitted a PR to [docs repository](https://github.com/blockscout/docs) to update the list of [env vars](https://github.com/blockscout/docs/blob/master/setup/env-variables/README.md) and I updated the version to `master` in the Version column. If I removed variable, I added it to [Deprecated ENV Variables](https://github.com/blockscout/docs/blob/master/setup/env-variables/deprecated-env-variables/README.md) page. After merging docs PR, changes will be reflected in these [pages](https://docs.blockscout.com/setup/env-variables).
- [ ] If I added new DB indices, I checked, that they are not redundant, with PGHero or other tools.
- [ ] If I added/removed chain type, I modified the Github CI matrix and PR labels accordingly.


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
  - Introduced benchmarking tools for measuring API endpoint performance.
  - Added support for aggregating transaction metrics within blocks, including fees and transaction counts.
- **Refactor**
  - Centralized transaction aggregation logic, simplifying how transaction metrics are accessed and rendered.
  - Improved cache update process by sanitizing data before storage to optimize memory and performance.
  - Enhanced caching behaviour to allow element sanitization before update.
  - Added helper function for consistent numeric to decimal conversions.
- **Bug Fixes**
  - Ensured precomputed transaction metrics are used throughout the application for consistency.
- **Chores**
  - Added benchmarking dependencies and support files for improved testing and performance analysis.
  - Updated translation source references to reflect codebase changes.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->